Rename controller class source and header files
