+++ app/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_executable(shell-app main.cpp PID_controller.cpp)
+add_executable(shell-app main.cpp controller.cpp)
 include_directories(
     ${CMAKE_SOURCE_DIR}/include
 )
