+++ test/CMakeLists.txt
@@ -4,7 +4,7 @@ add_executable(
     cpp-test
     main.cpp
     test.cpp
-    ../app/PID_controller.cpp
+    ../app/controller.cpp
 )
 
 target_include_directories(cpp-test PUBLIC ../vendor/googletest/googletest/include
